docker-compose structure done
