+++ .gitignore
@@ -1,3 +1,3 @@
 /venv/
-/.env
-/sip.conf
+/yeabook/.env
+/yeabook/sip.conf
