+++ docker-compose.yml
@@ -1,0 +1,22 @@
+version: '3'
+services:
+  yeabook:
+    build: ./yeabook
+    env_file:
+      - ./yeabook/.env
+    volumes:
+      - static_volume:/opt/yeabook/files/
+
+  nginx:
+    image: nginx:latest
+    volumes:
+      - ./nginx/nginx.conf:/etc/nginx/nginx.conf:ro
+      - ./nginx/configs:/etc/nginx/conf.d:ro
+      - static_volume:/var/www/yeabook/
+    depends_on:
+      - yeabook
+    ports:
+      - "80:80"
+
+volumes:
+  static_volume:
