+++ nginx/configs/site.conf
@@ -1,0 +1,19 @@
+server {
+    listen       80 default_server;
+    listen       [::]:80 default_server;
+    server_name  _;
+
+    root /opt/yeabook;
+
+    location /book/ {
+        autoindex on;
+        expires 90d;
+        root /var/www/;
+    }
+
+    error_page   404              /404.html;
+    error_page   500 502 503 504  /50x.html;
+    location = /50x.html {
+        root   html;
+    }
+}
