+++ nginx/nginx.conf
@@ -1,0 +1,37 @@
+worker_processes 1;
+
+events {
+    worker_connections 1024;
+}
+
+http {
+    include     mime.types;
+    log_format  main '$remote_addr - $remote_user [$time_local] "$request" '
+                      '$status $body_bytes_sent "$http_referer" '
+                      '"$http_user_agent" "$http_x_forwarded_for"';
+    sendfile    on;
+    tcp_nodelay on;
+    tcp_nopush  on;
+    client_max_body_size    200m;
+    server_tokens off;
+
+    gzip on;
+    gzip_comp_level 3;
+    gzip_min_length 1000;
+    gzip_types
+        text/plain
+        text/css
+        application/json
+        application/x-javascript
+        text/xml
+        text/javascript;
+
+    proxy_redirect     off;
+    proxy_set_header   Host             $host;
+    proxy_set_header   X-Real-IP        $remote_addr;
+    proxy_set_header   X-Forwarded-For  $proxy_add_x_forwarded_for;
+    include     conf.d/*.conf;
+
+    set_real_ip_from  192.168.1.0/24;
+    real_ip_header    X-Forwarded-For;
+}
